+++ app/app/component/db_func.py
@@ -5,7 +5,7 @@ import sys
 
 from flask import current_app
 
-from app.flaskr.models import ProjectModel, ProjectDeviceModel, ProjectPropertiesModel, ProjectUserModel
+from app.flaskr.sa.models import ProjectDeviceModel, ProjectPropertiesModel, ProjectUserModel
 from app.my_extensions import db
 from app.utils.database import model_to_dict
 
@@ -17,36 +17,6 @@ from app.component.db_op import do_tidb_exe,do_tidb_select
 from app.component.public_value import get_display_day
 from app.configs.export import write_to_log
 from app.configs import admin
-
-
-def insert_event(request_data):
-    project_model = request_data.to_project_model()
-
-    db.session.add(project_model)
-    # 事务
-    db.session.commit()
-    current_app.logger.debug(f'插入事件数据成功， 数据为: {model_to_dict(project_model)}')
-    return 1
-
-
-def insert_or_update_device(request_data):
-    request_data.set_other_properties()
-    count = insert_device_db(request_data)
-    current_app.logger.debug(f'插入或更新device={count}条')
-
-
-def insert_device_db(request_data, created_at=None, updated_at=None):
-
-    distinct_id = request_data.distinct_id
-    project_device_model_db = ProjectDeviceModel.query.filter_by(distinct_id=distinct_id).first()
-    if project_device_model_db is None:
-        project_device_model_db = request_data.to_project_device_model()
-        db.session.add(project_device_model_db)
-    else:   # 更新非空字段信息
-        request_data.update_project_device_model(project_device_model_db)
-
-    db.session.commit()
-    return 1
 
 
 def insert_user_db(project,distinct_id,lib,map_id,original_id,user_id,all_user_profile,update_params,created_at=None,updated_at=None):
@@ -62,46 +32,6 @@ def insert_user_db(project,distinct_id,lib,map_id,original_id,user_id,all_user_p
     key = {'distinct_id':distinct_id,'lib':lib,'map_id':map_id,'original_id':original_id,'user_id':user_id,'all_user_profile':all_user_profile,'created_at':timenow,'updated_at':timenow}
     result = do_tidb_exe(sql=sql, args=key)
     return result[1]
-
-
-def insert_properties(request_data):
-    table = request_data.project
-    lib = request_data.lib
-    remark = request_data.remark
-    event = request_data.event
-
-    ProjectPropertiesModel.__table__.name = f'{table}_properties'
-    project_properties_model_db = ProjectPropertiesModel.query.filter_by(event=event, lib=lib, remark=remark).first()
-    if project_properties_model_db is None:
-        project_properties_model_db = request_data.to_project_device_model()
-        db.session.add(project_properties_model_db)
-    else:   # 更新非空字段信息
-        request_data.update_project_properties_model(project_properties_model_db)
-
-    db.session.commit()
-    current_app.logger.debug(f'插入或更新事件属性数据成功， 数据为: {model_to_dict(project_properties_model_db)}')
-    return 1
-
-
-# def insert_user(project,data_decode,created_at=None):
-def insert_user(request_data):
-    table = request_data.project
-    distinct_id = request_data.distinct_id
-    lib = request_data.lib
-    map_id = request_data.map_id
-    original_id = request_data.original_id
-
-    ProjectUserModel.__table__.name = f'{table}_user'
-    project_user_model_db = ProjectUserModel.query.filter_by(distinct_id=distinct_id, lib=lib, map_id=map_id, original_id=original_id).first()
-    if project_user_model_db is None:
-        project_user_model_db = request_data.to_project_user_model()
-        db.session.add(project_user_model_db)
-    else:  # 更新非空字段信息
-        request_data.update_project_properties_model(project_user_model_db)
-
-    db.session.commit()
-    current_app.logger.debug(f'插入或更新用户属性数据成功， 数据为: {model_to_dict(project_user_model_db)}')
-    return 1
 
 
 def get_long_url_from_short(short_url):
@@ -981,16 +911,19 @@ def select_msg_type():
     result = do_tidb_select(sql=sql)
     return result
 
+
 def select_properties(project):
     sql="""select `event`,min(access_control_threshold) from {project}_properties where access_control_threshold is not null and access_control_threshold>0 GROUP BY `event`""".format(project=project)
     result = do_tidb_select(sql=sql)
     return result
+
 
 def insert_update_access_control_list(project,key,type_int,event,pv,date,hour):
     sql="""insert into `access_control` (`project`,`key`,`type`,`event`,`status`,`date`,`hour`,`pv`,`updated_at`) VALUES  (%(project)s,%(key)s,%(type)s,%(event)s,%(status)s,%(date)s,%(hour)s,%(pv)s,%(updated_at)s) ON DUPLICATE KEY UPDATE `updated_at`=%(updated_at)s,`pv` = `pv`+ %(pv)s;"""
     key = {'project':project,'key':key,'type':type_int,'event':event,'status':57,'date':date,'hour':hour,'pv':pv,'updated_at':int(time.time())}
     result = do_tidb_exe(sql=sql,args=key,retrycount=0)
     return result
+
 
 def query_access_control(key,project=None,type_id=None,event=None,date=None,hour=None,pv=None,query_hour=None,arr_mode=None):
     arr_mode = arr_mode if arr_mode else 'each'
