+++ app/app/configs/__init__.py
@@ -2,4 +2,8 @@
 # -*- coding: utf-8 -*-
 """
     ghost_sa配置目录.
+    default.py      默认配置
+    development.py 开发环境配置
+    test.py         测试环境配置
+    production.py   生产环境配置
 """
