+++ app/app/configs/development.py
@@ -24,6 +24,8 @@ class DevelopmentConfig(DefaultConfig):
 
     WTF_CSRF_ENABLED = False
 
+    # 开启kafka接收数据
+    USE_KAFKA = True
     # Kafka的Topic
     KAFKA_TOPIC = 'events-tracking_test_cero'
