+++ app/app/flaskr/__init__.py
@@ -1,5 +1,7 @@
 # !/usr/bin/python
 # -*- coding: utf-8 -*-
 """
-    request请求配置目录.
+    各蓝图模块目录
+    1. sa蓝图模块，提供鬼策入口模式
+    2. _tmp为动态模型目录，生成动态模型
 """
