+++ app/app/flaskr/_tmp/__init__.py
@@ -1,0 +1,7 @@
+# !/usr/bin/python
+# -*- coding: utf-8 -*-
+"""
+    动态生成临时模型目录.
+    此目录已加入.gitignore文件
+    程序中动态生成相应文件和模型，动态引入程序
+"""
