+++ app/app/flaskr/sa/__init__.py
@@ -1,0 +1,5 @@
+# !/usr/bin/python
+# -*- coding: utf-8 -*-
+"""
+    鬼策入口模块，此为sa.gif和sa入口处理核心模块
+"""
